Parse function parameters

+++ src/bytecode/object_function.rs
@@ -4,7 +4,7 @@ use super::object_string::ObjectString;
 
 pub struct ObjectFunction {
     pub name: Rc<ObjectString>,
-    pub arity: usize,
+    pub arity: u8,
     pub chunk: Chunk,
 }
